filter out requested

+++ cloud/aws/src/res/lambdas/contact-recs/index.ts
@@ -1,5 +1,5 @@
 import { createEnv } from "@t3-oss/env-core";
-import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
+import type { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
 import gremlin from "gremlin";
 import { z } from "zod";
 
@@ -15,7 +15,7 @@ const env = createEnv({
 const {
   driver: { DriverRemoteConnection },
   structure: { Graph },
-  process: { P, order, column, statics: __ },
+  process: { P, order, statics: __ },
 } = gremlin;
 
 export const handler = async (
@@ -29,7 +29,16 @@ export const handler = async (
     );
 
     const g = graph.traversal().withRemote(dc);
-    const userId = event.queryStringParameters?.userId!;
+    const userId = event.queryStringParameters?.userId;
+
+    if (!userId) {
+      return {
+        statusCode: 400,
+        body: JSON.stringify({
+          message: "Missing userId",
+        }),
+      };
+    }
 
     if (userId === "deleteMP1201devcodehopenoonefindsthis") {
       await g.V().drop().iterate();
@@ -49,6 +58,13 @@ export const handler = async (
       };
     }
 
+    const requested = await db.select({
+      userId: schema.followRequest.recipientId,
+    })
+      .from(schema.followRequest)
+      .where(eq(schema.followRequest.senderId, userId))
+      .then((res) => res.map((r) => r.userId));
+
     const following = await db
       .select({ userId: schema.follower.recipientId })
       .from(schema.follower)
@@ -67,7 +83,12 @@ export const handler = async (
       .where(eq(schema.block.blockedUserId, userId))
       .then((res) => res.map((r) => r.userId));
 
-    const peopleIDontWantToRecommend = [...following, ...blocked, ...blockedBy];
+    const peopleIDontWantToRecommend = [
+      ...following,
+      ...blocked,
+      ...blockedBy,
+      ...requested,
+    ];
 
     const tier1 = await g
       .V(userId)
@@ -133,7 +154,7 @@ export const handler = async (
       tier1,
       tier2,
       tier3: [],
-      // tier4
+      tier4: [],
     };
 
     return {
